+++ lib/features/posts/models/post_tabs/employee_post_dtos.dart
@@ -1,5 +1,5 @@
 import 'package:knowme_frontend/features/posts/models/contests_model.dart';
-import 'package:knowme_frontend/features/posts/models/basepost_model.dart';
+import 'package:knowme_frontend/features/posts/models/post_tabs/basepost_model.dart';
 
 // 채용 공고 모델
 class EmployeePost extends BasePost {
