+++ lib/features/posts/services/education_api_service.dart
@@ -1,5 +1,5 @@
 import '../../../shared/services/api_client.dart';
-import '../models/education_post_dtos.dart';
+import '../models/post_tabs/education_post_dtos.dart';
 import 'package:logger/logger.dart';
 
 /// 교육/강연 공고 API 서비스
